Factor TChannelPeersBase out of TChannelPeers
