+++ node/peers.js
@@ -22,8 +22,8 @@
 
 var inherits = require('util').inherits;
 var extend = require('xtend');
-var EventEmitter = require('./lib/event_emitter');
 
+var TChannelPeersBase = require('./peers_base.js');
 var TChannelPeer = require('./peer');
 var TChannelSelfPeer = require('./self_peer');
 
@@ -32,20 +32,15 @@ function TChannelPeers(channel, options) {
         return new TChannelPeers(channel, options);
     }
     var self = this;
-    EventEmitter.call(self);
-    self.allocPeerEvent = self.defineEvent('allocPeer');
+    TChannelPeersBase.call(self, channel, options);
 
-    self.channel = channel;
-    self.logger = self.channel.logger;
-    self.options = options || {};
+    self.allocPeerEvent = self.defineEvent('allocPeer');
     self.peerOptions = self.options.peerOptions || {};
     self.peerScoreThreshold = self.options.peerScoreThreshold || 0;
-    self._map = Object.create(null);
-    self._keys = [];
     self.selfPeer = null;
 }
 
-inherits(TChannelPeers, EventEmitter);
+inherits(TChannelPeers, TChannelPeersBase);
 
 TChannelPeers.prototype.close = function close(callback) {
     var self = this;
@@ -54,56 +49,25 @@ TChannelPeers.prototype.close = function close(callback) {
     if (self.selfPeer) {
         peers.push(self.selfPeer);
     }
-    var counter = peers.length + 1;
-    peers.forEach(function eachPeer(peer) {
-        peer.close(onClose);
-    });
-    self.clear();
-    onClose();
-
-    function onClose() {
-        if (--counter <= 0) {
-            if (counter < 0) {
-                self.logger.error('closed more peers than expected', {
-                    counter: counter
-                });
-            }
-            callback();
-        }
-    }
+    TChannelPeersBase.prototype.close.call(self, peers, callback);
 };
 
 TChannelPeers.prototype.sanitySweep = function sanitySweep() {
     var self = this;
 
-    var i = 0;
-    var conn = null;
-
     if (self.selfPeer) {
-        for (i = 0; i < self.selfPeer.connections.length; i++) {
-            conn = self.selfPeer.connections[i];
+        for (var i = 0; i < self.selfPeer.connections.length; i++) {
+            var conn = self.selfPeer.connections[i];
             conn.ops.sanitySweep();
         }
     }
-
-    var peers = self.values();
-    for (i = 0; i < peers.length; i++) {
-        var peer = peers[i];
-        for (var j = 0; j < peer.connections.length; j++) {
-            conn = peer.connections[j];
-            conn.ops.sanitySweep();
-        }
-    }
-};
-
-TChannelPeers.prototype.get = function get(hostPort) {
-    var self = this;
-    return self._map[hostPort] || null;
+    TChannelPeersBase.prototype.sanitySweep.call(self);
 };
 
 TChannelPeers.prototype.add = function add(hostPort, options) {
     /*eslint max-statements: [2, 25]*/
     var self = this;
+
     var peer = self._map[hostPort];
     if (peer) {
         return peer;
@@ -133,31 +97,6 @@ TChannelPeers.prototype.add = function add(hostPort, options) {
     return peer;
 };
 
-TChannelPeers.prototype.keys = function keys() {
-    var self = this;
-    return self._keys.slice();
-};
-
-TChannelPeers.prototype.values = function values() {
-    var self = this;
-    var keys = self._keys;
-    var ret = new Array(keys.length);
-    for (var i = 0; i < keys.length; i++) {
-        ret[i] = self._map[keys[i]];
-    }
-    return ret;
-};
-
-TChannelPeers.prototype.entries = function entries() {
-    var self = this;
-    var keys = self._keys;
-    var ret = new Array(keys.length);
-    for (var i = 0; i < keys.length; i++) {
-        ret[i] = [keys[i], self._map[keys[i]]];
-    }
-    return ret;
-};
-
 TChannelPeers.prototype.clear = function clear() {
     var self = this;
 
@@ -173,30 +112,22 @@ TChannelPeers.prototype.clear = function clear() {
     self._keys = [];
 };
 
-TChannelPeers.prototype.delete = function del(hostPort) {
+TChannelPeers.prototype._delete = function _del(peer) {
     var self = this;
-    var peer = self._map[hostPort];
-
-    if (!peer) {
-        return;
-    }
 
     if (self.channel.subChannels) {
         var names = Object.keys(self.channel.subChannels);
         for (var i = 0; i < names.length; i++) {
             var subChannel = self.channel.subChannels[names[i]];
-            subChannel.peers.delete(hostPort);
+            subChannel.peers.delete(peer.hostPort);
         }
     }
-    delete self._map[hostPort];
-    var index = self._keys.indexOf(hostPort);
+    delete self._map[peer.hostPort];
+    var index = self._keys.indexOf(peer.hostPort);
     self._keys.splice(index, 1);
-
-    return peer;
 };
 
-TChannelPeers.prototype.choosePeer =
-function choosePeer(req) {
+TChannelPeers.prototype.choosePeer = function choosePeer(req) {
     /*eslint complexity: [2, 15]*/
     var self = this;
